Serverbound/Clientbound rewrites.

+++ common/src/main/java/dev/greenhouseteam/orchestrate/song/Note.java
@@ -9,21 +9,21 @@ import net.minecraft.network.codec.StreamCodec;
 
 import java.util.Objects;
 
-public record Note(KeyWithOctave key, float volume, float startTime, float duration) {
+public record Note(KeyWithOctave key, float volume, int startTime, int duration) {
     public static final Codec<Note> CODEC = RecordCodecBuilder.create(inst -> inst.group(
             KeyWithOctave.CODEC.fieldOf("key").forGetter(Note::key),
             Codec.FLOAT.fieldOf("volume").forGetter(Note::volume),
-            Codec.FLOAT.fieldOf("start_time").forGetter(Note::startTime),
-            Codec.FLOAT.fieldOf("duration").forGetter(Note::duration)
+            Codec.INT.fieldOf("start_time").forGetter(Note::startTime),
+            Codec.INT.fieldOf("duration").forGetter(Note::duration)
     ).apply(inst, Note::new));
     public static final StreamCodec<RegistryFriendlyByteBuf, Note> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.fromCodec(KeyWithOctave.CODEC),
             Note::key,
             ByteBufCodecs.FLOAT,
             Note::volume,
-            ByteBufCodecs.FLOAT,
+            ByteBufCodecs.INT,
             Note::startTime,
-            ByteBufCodecs.FLOAT,
+            ByteBufCodecs.INT,
             Note::duration,
             Note::new
     );
